+++ src/background/state/application/toolbar_items.rs
@@ -28,7 +28,7 @@ impl FullState {
             center: vec![ToolbarItem2::Plugin("@default/date".into())],
             right: vec![
                 ToolbarItem2::Plugin("@seelen/tb-system-tray".into()),
-                ToolbarItem2::Plugin("@default/keyboard".into()),
+                ToolbarItem2::Plugin("@seelen/tb-keyboard-selector".into()),
                 ToolbarItem2::Plugin("@seelen/tb-bluetooth-popup".into()),
                 ToolbarItem2::Plugin("@default/network".into()),
                 ToolbarItem2::Plugin("@default/media".into()),